Add additional editor events
